+++ src/components/Buttons/InterruptTrainExperimentButton/InterruptTrainExperimentButton.md
@@ -1,0 +1,20 @@
+**Default**
+
+```js
+<InterruptTrainExperimentButton handleClick={() => alert('Clicked')} />
+```
+
+**Loading**
+
+```js
+<InterruptTrainExperimentButton
+  handleClick={() => alert('Clicked')}
+  deleteExperimentRunning
+/>
+```
+
+**Disabled**
+
+```js
+<InterruptTrainExperimentButton handleClick={() => alert('Clicked')} disabled />
+```
