+++ src/components/Buttons/RemoveOperatorButton/RemoveOperatorButton.md
@@ -1,0 +1,17 @@
+**Default**
+
+```js
+<RemoveOperatorButton handleClick={() => alert('Clicked')} />
+```
+
+**Loading**
+
+```js
+<RemoveOperatorButton handleClick={() => alert('Clicked')} loading />
+```
+
+**Loading**
+
+```js
+<RemoveOperatorButton handleClick={() => alert('Clicked')} disabled />
+```
