+++ src/components/Buttons/TrainExperimentButton/TrainExperimentButton.md
@@ -1,0 +1,11 @@
+**Default**
+
+```js
+<TrainExperimentButton handleClick={() => alert('Clicked')} />
+```
+
+**Disabled**
+
+```js
+<TrainExperimentButton handleClick={() => alert('Clicked')} disabled />
+```
