+++ src/components/Buttons/UploadInferenceTestButton/UploadInferenceTestButton.component.jsx
@@ -1,0 +1,52 @@
+// CORE LIBS
+import React from 'react';
+import PropTypes from 'prop-types';
+
+// UI LIBS
+import { FundOutlined } from '@ant-design/icons';
+import { Button, Tooltip, Upload } from 'antd';
+
+/**
+ * Upload Inference Test Button.
+ * This component is responsible for displaying upload inference test button.
+ */
+const UploadInferenceTestButton = ({ disabled, handleUpload }) => {
+  // upload props
+  const props = {
+    name: 'file',
+    showUploadList: false,
+    accept: '*',
+  };
+
+  // RENDER
+  return (
+    // upload component
+    <Upload
+      beforeUpload={(file) => {
+        handleUpload(file);
+      }}
+      disabled={disabled}
+      {...props}
+    >
+      <Tooltip placement='bottom' title='Testar o fluxo'>
+        <Button
+          disabled={disabled}
+          size='large'
+          style={{ padding: 0 }}
+          type='link'
+        >
+          <FundOutlined />
+        </Button>
+      </Tooltip>
+    </Upload>
+  );
+};
+
+// PROP TYPES
+UploadInferenceTestButton.propTypes = {
+  /** upload inference test button upload handle */
+  handleUpload: PropTypes.func.isRequired,
+};
+
+// EXPORT
+export default UploadInferenceTestButton;
