+++ src/components/Buttons/UploadInferenceTestButton/UploadInferenceTestButton.md
@@ -1,0 +1,11 @@
+**Default**
+
+```js
+<UploadInferenceTestButton handleUpload={() => alert('Upload')} />
+```
+
+**Disabled**
+
+```js
+<UploadInferenceTestButton handleUpload={() => alert('Upload')} disabled />
+```
